move ansible help parsing to parse_help.py
